Crash if chat is already deleted (#321)

+++ src/Store/make-in-memory-store.ts
@@ -227,7 +227,9 @@ export default (
 		})
 		ev.on('chats.delete', deletions => {
 			for(const item of deletions) {
-				chats.deleteById(item)
+				if(chats.get(item)) {
+					chats.deleteById(item)
+				}
 			}
 		})
 		ev.on('messages.upsert', ({ messages: newMessages, type }) => {
